Add a Snabb program to construct a complete L2VPN endpoint
program/l2vpn creates a set of apps that implements an arbitrary
number of VPLS instances sharing a single uplink to the Internet from
a single configuration statement.  A VPLS instance is uniquely
identified by an IPv6 address that serves as the destination for all
pseudowires associated with it.  A VPLS can have any number of
customer ports ("attachement circuits").  A multi-point VPLS uses an
instance of a flooding or learning bridge connected to a full mesh of
pseudowires.